+++ app/src/main/java/com/example/npod/data/nasa/NasaApi.kt
@@ -1,7 +1,7 @@
-package com.example.npod.api
+package com.example.npod.data.nasa
 
-import com.example.npod.data.dto.PhotoMarsListDTO
-import com.example.npod.data.dto.PictureOfTheDayDTO
+import com.example.npod.data.nasa.photo.PhotoMarsListDTO
+import com.example.npod.data.nasa.pod.PictureOfTheDayDTO
 import retrofit2.http.GET
 import retrofit2.http.Query
