+++ app/src/main/java/com/example/npod/data/nasa/NasaRepositoryImpl.kt
@@ -1,11 +1,11 @@
-package com.example.npod.data
+package com.example.npod.data.nasa
 
 import com.example.npod.BuildConfig
-import com.example.npod.data.dto.toPhotoMars
-import com.example.npod.data.dto.toPictureOfTheDay
-import com.example.npod.domain.models.PhotoMars
+import com.example.npod.data.nasa.photo.toPhotoMars
+import com.example.npod.data.nasa.pod.toPictureOfTheDay
+import com.example.npod.domain.PhotoMars
 import com.example.npod.domain.NasaRepository
-import com.example.npod.domain.models.PictureOfTheDay
+import com.example.npod.domain.PictureOfTheDay
 
 class NasaRepositoryImpl : NasaRepository {
     override suspend fun getPictureOfTheDay(date: String): PictureOfTheDay {
@@ -15,8 +15,7 @@ class NasaRepositoryImpl : NasaRepository {
     }
 
     override suspend fun getMarsPhoto(date: String): List<PhotoMars> {
-        return Retrofit
-            .getNasaApi()
+        return Retrofit.getNasaApi()
             .getPhotosMarsList(date, BuildConfig.API_KEY)
             .photos.map { it.toPhotoMars() }
     }
