+++ app/src/main/java/com/example/npod/data/nasa/photo/PhotoMarsListDTO.kt
@@ -1,6 +1,6 @@
-package com.example.npod.data.dto
+package com.example.npod.data.nasa.photo
 
-import com.example.npod.domain.models.PhotoMars
+import com.example.npod.domain.PhotoMars
 import com.google.gson.annotations.SerializedName
 
 data class PhotoMarsListDTO(
