+++ app/src/main/java/com/example/npod/data/nasa/photo/PhotoState.kt
@@ -1,6 +1,6 @@
-package com.example.npod.data
+package com.example.npod.data.nasa.photo
 
-import com.example.npod.domain.models.PhotoMars
+import com.example.npod.domain.PhotoMars
 
 sealed class PhotoState {
     data class Success(val photos: List<PhotoMars>) : PhotoState()
