+++ app/src/main/java/com/example/npod/ui/screens/MainViewModel.kt
@@ -2,7 +2,7 @@ package com.example.npod.ui.screens
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.npod.data.PictureState
+import com.example.npod.data.nasa.pod.PictureState
 import com.example.npod.domain.NasaRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
