+++ app/src/main/java/com/example/npod/ui/screens/photos/PhotosViewModel.kt
@@ -2,7 +2,7 @@ package com.example.npod.ui.screens.photos
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.npod.data.PhotoState
+import com.example.npod.data.nasa.photo.PhotoState
 import com.example.npod.domain.NasaRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
